backend/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi[standard]>=0.116.1",
    "ollama>=0.5.3",
    "openai>=1.99.9",
    "retry>=0.9.2",
    "tenacity>=9.1.2",
]
